Added Easy AI, few other changes
The main thing here is the Easy level AI. It simply returns drop as a random integer between 1 and 7.  Other than that, I moved the int drop from local within a function to a global variable, since it will be used by all the AI functions.  Another change was adding a clear to the draw board, which overall cleans up the console output a bunch.